requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
flask==3.0.2 